formatted code with black and rearranged folder structure removing extra code for other fixes
