+++ environment.yml
@@ -11,5 +11,6 @@ dependencies:
   - matplotlib=3.1.3
   - seaborn=0.10.0
   - pip=20.0.2
+ 
   - pip:
     - plotnine==0.6.0
